Specify Wrapper Query Directly
Previously the wrapper component specified the query like any other
component through om/IQuery. The problem with this is that all child
components of the wrapper don't have a valid om/full-query anymore
because there queries are not part of the wrapper query.

At the end not the wrapper has the query, the root component has the
query alongside the routing queries. We could rename the query in
something like static-query. This way we have the routing queries and
one static query and also static-props.